refactor: catalog - 멤버 레플리카 패키지 및 네이밍 개선
- CatalogMember → Member로 클래스명 변경
- replica.member 패키지로 이동
