+++ bc/catalog/src/main/java/app/giftify/product/application/ProductGetUseCase.java
@@ -1,10 +1,10 @@
 package app.giftify.product.application;
 
-import app.giftify.catalogmember.adapter.outbound.jpa.CatalogMemberRepository;
-import app.giftify.catalogmember.core.domain.CatalogMember;
 import app.giftify.product.adapter.inbound.ProductDto;
 import app.giftify.product.domain.Product;
 import app.giftify.product.domain.exception.ProductException;
+import app.giftify.replica.member.Member;
+import app.giftify.replica.member.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -16,14 +16,14 @@ import static app.giftify.product.domain.exception.ProductErrorCode.SELLER_NOT_F
 @RequiredArgsConstructor
 public class ProductGetUseCase {
     private final ProductSupport productSupport;
-    private final CatalogMemberRepository catalogMemberRepository;
+    private final MemberRepository memberRepository;
 
     /**
      * 상품 단건 조회
      */
     public ProductDto getProduct(Long productId) {
         Product product = productSupport.findById(productId); // 상품이 없으면 404
-        CatalogMember seller = catalogMemberRepository.findById(product.getSellerId())
+        Member seller = memberRepository.findById(product.getSellerId())
                 .orElseThrow(() -> new ProductException(SELLER_NOT_FOUND)); // 판매자가 없으면 404 (todo 판매자 탈퇴하면?)
 
         // 판매중이 아닌 상품
