+++ bc/catalog/src/main/java/app/giftify/replica/member/MemberRepository.java
@@ -1,0 +1,6 @@
+package app.giftify.replica.member;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+}
